[python-frontend] extract loop transformations into loop mixin and integrate preprocessor inheritance
